# models/post_package.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
)

// PostPackage Package
//
// Package object
// swagger:model PostPackage
type PostPackage struct {

	// blob
	//
	// Package blob: a tar.gz that is b64
	Blob string `json:"blob,omitempty"`

	// package-name
	//
	// Package name
	Package string `json:"package,omitempty"`

	// package-version
	//
	// Package version
	Release string `json:"release,omitempty"`
}

// Validate validates this post package
func (m *PostPackage) Validate(formats strfmt.Registry) error {
	var res []error

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

// MarshalBinary interface implementation
func (m *PostPackage) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *PostPackage) UnmarshalBinary(b []byte) error {
	var res PostPackage
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}