# swagger.yml
swagger: "2.0"
info:
  description: |
    Ksonnet Registry API documentation
  version: "0.1.0"
  title: Ksonnet Registry API
  contact:
    name: bryan@heptio.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: localhost:5000
basePath: /
schemes:
  - http
paths:
...
  /api/v1/packages/{namespace}/{package}:
    get:
      tags:
        - package
      summary: List all releases for a package
      description: ""
      operationId: showPackageReleases
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: namespace
          description: namespace
          required: true
          type: string
        - in: path
          name: package
          description: package name
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PackageManifest"
        "404":
          description: Package not found
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: Not authorized to read the package
          schema:
            $ref: "#/definitions/Error"

    post:
      tags:
        - package
      summary: Push new package release to the registry
      description: ""
      operationId: createPackage
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: namespace
          description: Namespace for package
          type: string
          required: true

        - in: path
          name: package
          description: Package for release
          type: string
          required: true

        - in: query
          name: force
          description: Force push the release (if allowed)
          required: false
          type: boolean
          default: false

        - in: body
          name: body
          description: Package object to be added to the registry
          required: true
          schema:
            $ref: "#/definitions/PostPackage"
      responses:
        "201":
          description: successful operation
          schema:
            $ref: "#/definitions/Package"
        "404":
          description: Package not found (if force=true)
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: Package already exists
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: Not authorized to create the package
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: Bad version or name format
          schema:
            $ref: "#/definitions/Error"
...
  /api/v1/packages/{namespace}/{package}/{release}/pull:
    get:
      tags:
        - package
        - blobs
      summary: Download the package
      description: ""
      operationId: pullPackage
      consumes:
        - application/json
      produces:
        - 'application/x-gzip'
      parameters:
        - in: path
          name: namespace
          description: namespace
          required: true
          type: string

        - in: path
          name: package
          description: package name
          required: true
          type: string

        - in: path
          name: release
          description: release name
          required: true
          type: string

      responses:
        "200":
          description: successful operation
          schema:
            type: file
            title: package-targz
        "404":
          description: Package not found
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: Not authorized to read the package
          schema:
            $ref: "#/definitions/Error"
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        description: http status code
        title: http-code
      message:
        type: string
        description: "error message"
      details:
        type: object
        description: error extra data
  PackageManifest:
    type: array
    title: manifests
    description: manifests
    items:
      $ref: "#/definitions/Manifest"
  PartDescriptor:
    type: object
    description: part descriptor
    title: descriptor
    properties:
      digest:
        type: string
        title: digest
        description: content digest
      size:
        type: integer
        format: int64
        title: content-size
        description: blob size
      dependencies:
        type: array
        title: dependencies
        description: dependencies
        items:
          $ref: "#/definitions/Dependency"
  Dependency:
    type: object
    title: dependency
    properties:
      name:
        type: string
        description: Dependency name
      constraint:
        type: string
        description: Dependency constraint
  Manifest:
    type: object
    title: manifest
    properties:
      created_at:
        type: string
        format: date-time
        description: creation data
        title: created-at
      metadata:
        type: object
        description: KeyValue object to add complementary and format specific information
        title: metadata
      package:
        type: string
        description: package name
        title: package-name
      release:
        type: string
        description: release name
        title: release-name
      content:
        $ref: "#/definitions/PartDescriptor"
...
  Package:
    type: object
    title: Package
    description: Package object
    properties:
      created_at:
        type: string
        format: date-time
        title: created_at
        description: Package creation date
      content:
        $ref: "#/definitions/Manifest"
      package:
        type: string
        title: package-name
        description: Package name
      release:
        type: string
        title: package-release
        description: Package release
...
  PostPackage:
    type: object
    title: Package
    description: Package object
    properties:
      blob:
        type: string
        title: blob
        description: "Package blob: a tar.gz that is b64"
      package:
        type: string
        title: package-name
        description: Package name
      release:
        type: string
        title: package-version
        description: Package version